+++ src/ts/component/util/graph.tsx
@@ -4,9 +4,8 @@ import $ from 'jquery';
 import * as d3 from 'd3';
 import { observer } from 'mobx-react';
 import { PreviewDefault } from 'Component';
-import { I, UtilCommon, UtilObject, UtilSmile, UtilFile, translate, Relation, analytics } from 'Lib';
+import { I, UtilCommon, UtilObject, UtilSmile, UtilGraph, translate, analytics } from 'Lib';
 import { commonStore } from 'Store';
-import Colors from 'json/colors.json';
 
 interface Props {
 	isPopup?: boolean;
@@ -16,11 +15,6 @@ interface Props {
 	onContextMenu?: (id: string, param: any) => void;
 	onContextSpaceClick?: (param: any, data: any) => void;
 	onSelect?: (id: string, related?: string[]) => void;
-};
-
-const bgColor = {
-	'': '#000',
-	dark: '#fff',
 };
 
 const Graph = observer(class Graph extends React.Component<Props> {
@@ -166,7 +160,7 @@ const Graph = observer(class Graph extends React.Component<Props> {
 	nodeMapper (d: any) {
 		d.layout = Number(d.layout) || 0;
 		d.radius = 4;
-		d.src = this.imageSrc(d);
+		d.src = UtilGraph.imageSrc(d);
 
 		if (d.layout == I.ObjectLayout.Note) {
 			d.name = d.snippet || translate('commonEmpty');
@@ -363,110 +357,6 @@ const Graph = observer(class Graph extends React.Component<Props> {
 			};
 
 		};
-	};
-
-	imageSrc (d: any) {
-		let src = '';
-
-		switch (d.layout) {
-			case I.ObjectLayout.Relation: {
-				src = `img/icon/relation/big/${Relation.typeName(d.relationFormat)}.svg`;
-				break;
-			};
-
-			case I.ObjectLayout.Task: {
-				src = `img/icon/graph/task.svg`;
-				break;
-			};
-
-			case I.ObjectLayout.File: {
-				src = `img/icon/file/${UtilFile.icon(d)}.svg`;
-				break;
-			};
-
-			case I.ObjectLayout.Image: {
-				if (d.id) {
-					src = commonStore.imageUrl(d.id, 100);
-				} else {
-					src = `img/icon/file/${UtilFile.icon(d)}.svg`;
-				};
-				break;
-			};
-				
-			case I.ObjectLayout.Human: {
-				if (d.iconImage) {
-					src = commonStore.imageUrl(d.iconImage, 100);
-				};
-				break;
-			};
-
-			case I.ObjectLayout.Note: {
-				break;
-			};
-
-			case I.ObjectLayout.Bookmark: {
-				if (d.iconImage) {
-					src = commonStore.imageUrl(d.iconImage, 100);
-				};
-				break;
-			};
-				
-			default: {
-				if (d.iconImage) {
-					src = commonStore.imageUrl(d.iconImage, 100);
-				} else
-				if (d.iconEmoji) {
-					const data = UtilSmile.data(d.iconEmoji);
-					if (data) {
-						src = UtilSmile.srcFromColons(data.colons, data.skin);
-					};
-					src = src.replace(/^.\//, '');
-				} else
-				if (d.iconOption) {
-					src = this.gradientIcon(d.iconOption, true);
-				};
-				break;
-			};
-		};
-
-		return src;
-	};
-
-	gradientIcon (iconOption: number, small?: boolean) {
-		const option: any = Colors.gradientIcons.options[iconOption - 1];
-		if (!option) {
-			return;
-		};
-
-		const theme = commonStore.getThemeClass();
-		const { from, to } = option.colors;
-		const canvas = document.createElement('canvas');
-		const ctx = canvas.getContext('2d');
-		const w = 160;
-		const r = w / 2;
-		const fillW = small ? w * 0.7 : w;
-		const fillR = fillW / 2;
-		const steps = option.steps || Colors.gradientIcons.common.steps;
-		const step0 = UtilCommon.getPercentage(fillR, steps.from * 100);
-		const step1 = UtilCommon.getPercentage(fillR, steps.to * 100);
-		const grd = ctx.createRadialGradient(r, r, step0, r, r, step1);
-
-		canvas.width = w;
-		canvas.height = w;
-		grd.addColorStop(0, from);
-		grd.addColorStop(1, to);
-
-		if (small) {
-			ctx.fillStyle = bgColor[theme];
-			ctx.fillRect(0, 0, w, w);
-		};
-
-		ctx.fillStyle = grd;
-		ctx.beginPath();
-		ctx.arc(r, r, fillR, 0, 2 * Math.PI);
-		ctx.fill();
-
-		return canvas.toDataURL();
 	};
 
 	setRootId (id: string) {
